src/main/webapp/app/entities/expenseService/employee-2/delete/employee-2-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { Employee2Service } from '../service/employee-2.service';

import { Employee2DeleteDialogComponent } from './employee-2-delete-dialog.component';

describe('Employee2 Management Delete Component', () => {
  let comp: Employee2DeleteDialogComponent;
  let fixture: ComponentFixture<Employee2DeleteDialogComponent>;
  let service: Employee2Service;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, Employee2DeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(Employee2DeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(Employee2DeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(Employee2Service);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('9fec3727-3421-4967-b213-ba36557ca194');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('9fec3727-3421-4967-b213-ba36557ca194');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
